更新机械结构设计，更新部分demo，优化部分api，修复部分bug
0. 更新K210端代码
1. 更新机械结构设计
2. 更新PWM库设计

+++ STM32/Src/App/ArmControl/LiftingPlatform.c
@@ -24,7 +24,7 @@ static PositionPID_t xAxisPID = { 0 };
 static PositionPID_t yAxisPID = { 0 };
 #endif
 
-static float currentZ_AxisHeight = 0.0, currentAxisLength = 0.0;
+static uint16_t currentZ_AxisHeight = 0.0, currentAxisLength = 0.0;
 
 /**
  * @brief: Init arm control of the robot.
